// src/app/shared/models/iframely-response.ts
export interface IframelyResponse {
  url: string;
  type: string;
  version: string;
  title: string;
  author: string;
  author_url: string;
  provider_name: string;
  description: string;
  thumbnail_url: string;
  thumbnail_width: number;
  thumbnail_height: number;
  html: string;
  cache_age: number;
  error: string;
}